moved eventer to service package top level
