+++ main.go
@@ -19,9 +19,9 @@ import (
 	"github.com/giantswarm/draughtsman/service"
 	"github.com/giantswarm/draughtsman/service/configurer/configmap"
 	"github.com/giantswarm/draughtsman/service/deployer"
-	"github.com/giantswarm/draughtsman/service/deployer/eventer/github"
 	"github.com/giantswarm/draughtsman/service/deployer/installer/helm"
 	slacknotifier "github.com/giantswarm/draughtsman/service/deployer/notifier/slack"
+	"github.com/giantswarm/draughtsman/service/eventer/github"
 	slackspec "github.com/giantswarm/draughtsman/service/slack"
 )
