+++ service/deployer/deployer.go
@@ -8,12 +8,12 @@ import (
 	micrologger "github.com/giantswarm/microkit/logger"
 
 	"github.com/giantswarm/draughtsman/flag"
-	"github.com/giantswarm/draughtsman/service/deployer/eventer"
-	eventerspec "github.com/giantswarm/draughtsman/service/deployer/eventer/spec"
 	"github.com/giantswarm/draughtsman/service/deployer/installer"
 	installerspec "github.com/giantswarm/draughtsman/service/deployer/installer/spec"
 	"github.com/giantswarm/draughtsman/service/deployer/notifier"
 	notifierspec "github.com/giantswarm/draughtsman/service/deployer/notifier/spec"
+	"github.com/giantswarm/draughtsman/service/eventer"
+	eventerspec "github.com/giantswarm/draughtsman/service/eventer/spec"
 	httpspec "github.com/giantswarm/draughtsman/service/http"
 	slackspec "github.com/giantswarm/draughtsman/service/slack"
 )
