+++ service/deployer/notifier/spec/spec.go
@@ -1,7 +1,7 @@
 package spec
 
 import (
-	"github.com/giantswarm/draughtsman/service/deployer/eventer/spec"
+	"github.com/giantswarm/draughtsman/service/eventer/spec"
 )
 
 // NotifierType represents the type of Notifier to configure.
